Added tech stack to projects page
This required me to delve into the [Jekyll data files](http://jekyllrb.com/docs/datafiles/)
which provide the ability to save shared data that can be accessed
sitewide. This provides me with the ability to avoid repetition, and
instead refer to a single item in each of the project pages' front
matter.