Display logo/image alongside project description
As well as displaying the logo, make our title and description a bit
larger, and fill the space a bit better.

Currently fit the image in a 40% split; this seems to work well for
some images, and not so well for others. It could be that we need to
adjust this, or provide the ability to grow the image on click