Integrate Jekyll and Gulp optimisations
In order to streamline the process for the site, we want to only need
to run one single command. As part of this, we want `gulp` to perform a
Jekyll build, and then optimise the rest of the site.

Additionally, we want to be able to serve the site on a local port so
we can test changes and see updates more easily. This also requires a
`watch` job which will poll for changes and regenerate the site.