Pull the latest image if current branch doesn't have one
In the case that we're building on a new branch, we're going to be doing
a completely fresh build, which is going to be slow. Instead, if we try
and pull whatever the latest image is, we'll hopefully have at least a
couple of layers which can be used as a cache.