Add warning when Tech Stack items aren't set
Add a fake link and big red text to indicate when we've got Tech Stack
items that aren't set. This is primarily because if they're unset
currently, they'll simply silently fail, and output empty `<li />` which
is far from ideal. This way, the fake link will cause `htmlproofer` to
fail in our CI pipeline.