Make collections plugin even more generic
In order to reduce the duplication even more, we now have a
`CollectionPageGenerator` which provides a set of overrideable
attributes that define what to generate.

One issue I had with this was getting the actual i.e. `tags` out of the
site. A workaround is using the `post_attr_hash` method that's exposed
by the `Jekyll::Site` class.

Another issue is that I couldn't get the instance variables working.
Once I found out about the `initialize` method, this was easily sorted.

Closes #66.