Interpolate value of `collectionLayoutHtml`
Instead of duplicating the filename around, let's interpolate it from
the `@collectionLayoutKey` variable.

Note that we have to create the helper setter method `set_collectionLayout`
such that we can call it at a later point - for instance, as
we're doing this within the superclass's `initialize` method. If we were
to override the value of `collectionLayoutKey`, it'd still refer to the
old one, as the interpolation is only done in the superclass.

As part of #66.