Default posts to using TOC
As found when writing the Why-Gitlab article, it turns out that the
plugin requires the presence of the `page.toc` attribute on each post.
Therefore, we can add this into the defaults[1] such that it will always
be there for all pages, and then we don't need the attribute anywhere.

[1]: https://github.com/jekyll/jekyll/issues/3046#issuecomment-177368881