Don't implicitly reference `post` variable
This was previously working in each of the pages because it was
implicitly able to get the `post.` variable.

Additionally, it fixes the fact that our post previews don't work, due
to the `post` variable not being available in the `read-time` snippet
(due to it being exposed as `page`).

As part of #117.