Improve readability of inline code snippets
As part of #118, try removing the different colouring of the inline code
snippets, so we still have the obvious backticks around the snippets,
but they aren't made less readable by the use of colour as well.

The colours were originally from before the snippets had backticks
around them, so was actually useful back then.