Deploy `builder` image to the server
As we don't tag the images as a top-level image until they hit `master`
and are released, we need to reference the `builder` image instead.

However, this would need to be reverted in #106, as at that point, the
builder will no longer be required - we'll have the base HTML to deploy,
which can be the top-level image.

As part of #64.