Build HTML-only container for generated site
As seen in #106, we want to generate an HTML-only image, such that the
pull on the other end is incredibly quick, as it only requires the HTML,
not the full builder image (i.e. 400MB down to 5MB).

Closes #131, as we push the generated content up to the top-level image
namespace.

We build the `built` image in the same method that we perform a
deployment currently - by copying the files out of the `builder`.