Display date(s) for talks
It's worth having the dates displayed for the talks, so readers can see
what events I have spoken at before, and where they can come to see the
talk in the future.

This is a first pass, so does not handle cases where either previous or
future events are not set, and therefore there is an empty list
displayed. Nor will any singular/plural distinctions (i.e. `event(s)`).

Additionally, update the schema for the talk events to make more sense.

As part of #112 and #111.