Ignore Git worktrees
Worktrees are a great feature that it's important we don't pull in
massive trees of unrelated content (which may duplicate existing posts
in the site, or break the build if a worktree doesn't conform to Jekyll
standards. Additionally, we have the bonus of our Docker build context
being smaller, as it won't need to pull all the worktrees into building
the image.