Increase width of content on page
The original `50rem` was actually a little narrow, and wastes a lot of
horizontal space on monitors. Increasing this slightly to `75rem` makes
for much more comfortable reading.

Note that we have to do the same for our code blocks, otherwise they're
indented oddly compared to the rest of our content.