Use file metadata to generate `lastmod`
As part of #200, we are generating a `sitemap.xml`. However, without the
`jekyll-last-modified-at` gem, we'll get a `lastmod` of the current
date, which doesn't correspond to the actual last modification date of
the item in our collections. Instead, if we use the file's last
modification date, this corresponds with the actual Git history of the
file.