Generate "read time" SEO tags for sharing
It's much nicer to be able to share posts and have a recipient know how
long the post will (roughly) take to read before they click through.

To do this, we take the advice from [0] and create a new property that
only exists on posts to denote when to show the "estimated reading
time".

Closes #87.

[0]: http://conductofcode.io/post/social-meta-tags-with-jekyll/