Remove `site.url` references where possible
We don't want to use `site.url` anywhere, especially as it's now been
configured as part of #160 to help with i.e. canonical URLs for SEO.

However, there are a number of places that actually already had the URLs
referenced, so these links would break local development, which means
they needed to be removed.