Fix: Mark prod site as an `internal domain` for HTML proofer
This is required to allow us to push up new pages on the site, as the
canonical url will always point to the prod URL. This means that every
time we push up a new site, we'll test the built output, and find that
actually, there isn't a page there, because it hasn't yet got there.

The way around this is to set `jvt.me` as one of the `internal_domains`
which ensures that HTMLProofer will not follow links with this there.
However, this _may_ open us up to further link breakage in the future.