Bake Capistrano config into deploy image
Related to #146, we don't need to redownload the whole site and its
history in our pipeline, just the Capistrano-specific config. This
should be baked into our deploy image, so we don't need to redownload
the sources on a deploy.

Note that our deploys now need to be out of the `/site` directory, as
that's where all the configuration is placed, and GitLab doesn't follow
Docker's` WORKDIR` directive.

Closes #150.