Set `JEKYLL_ENV` when building with `--production`
Although we're currently not using the `jekyll.environment` variables
within the site, we may as well have it there so it works, especially as
we've been passing the `--production` flag for a while without it doing
anything!

Note that we need to take the existing `process.env` so we can i.e.
inherit the existing `$PATH`, as well as other required environment
variables, and then set our own for `JEKYLL_ENV`.

Closes #199.