Display 10 posts on the front page
Three posts isn't very helpful, as it's quite a short list. Instead,
show 10, which goes back a good deal of time to show some variety, as
well as make it so a viewer doesn't necessarily have to click through to
`/posts/` to see if any of the other posts interest them.