Pipeline check that images have been minified
When running it locally we can find that actually, it hasn't been
correctly minified. This is pretty annoying, and will at least stop an
un-minified image from hitting the `master` branch.

Run this before checking links are well-formed via htmlproofer, as this
task runs much quicker and can provide us with an easier win, earlier.