Increase page `max-width`
This firstly helps increase the amount of content on the page, therefore
reducing use of vertical space, as more is in the horizontal space.

Secondly, this allows the social links to be placed in line with the
title + links, meaning on a wide enough screen, there is further
vertical space savings.

As part of #233.