Only display "last updated" date if different to publish date
As per Anna's comment, this should only be displayed if it's different
to the publish date, otherwise it's pretty useless information.

In the case that the two dates are the same, we don't need to show the
commit link, as it's not important what the change was (as it'd be the
publishing of the post).

Closes #259.