Add support for canonical URLs
For cases where I've imported articles from elsewhere, I need to be able
to support canonical URL linking.

Not only do we want to add the HTML metadata, but also output the
canonical URL to the user, so they're aware it is not the original
version.

This also needs to be added to our Kwalify schema, otherwise CI will
fail with invalid post metadata.

Closes #292.