Use same permalinks structure as existing site
Note that due to the way that Hugo builds the URL for pages from the
title's text, we need to specify the `slug` manually when importing our
content.

I don't want *any* breaking links as part of this change, so this is
really quite important! I've got some great backlinks to the site to
preserve.