Don't validate project/talk schema in pipeline
It's a pain to validate them while we're only partly rendering them in
the Hugo site (until we've migrated them to a microservice-site style
content type) so let's ignore it for now, and we'll pick this up soon.

These don't get edited that often that they need this to be here and
working.