Add initial Hugo archetype
This seems to work nicely enough that it generates a:

- `draft`
- with the correct `date` and `slug`, given the filename
- defaulting to the correct licenses

This isn't ideal, as it expects there to _always_ be a date in the
filename, which wasn't true before when creating drafts, so I'll have to
play around with it and see if it works.

And the generated file also matches the Kwalify schema!