Pull the extended version of Hugo
The extended version is required to build the SASS within our theme.

If we don't use the extended version, we're unable to build it and
receive a runtime error.

The extended Hugo doesn't seem to work on Alpine, so we need to pull it
on i.e. Debian, as we can then take advantage of the pre-packaged
installer.