Replace Jekyll markdown/code `include`s inline
Unfortunately Hugo's `partialCached` doesn't let you include non-HTML
codes, so instead of hacking around it, I'll just re-inline these large
includes of content.

Primarily the reason to move them out was to make sure that the posts
didn't break down markdown highlighting in Vim - which is a hack, and
should be resolved at Vim's end, not worked around in this.