Verify that Webmention endpoint is well formed
To ensure that I don't accidentally break my Webmention discovery in the
future, I want to use my newfound power with HTMLProofer's custom
`Check`s to catch future regressions in the metadata.

Note that we only check this on the home page as that's the important
place that a client would be looking it up - other pages may be valid,
but this is the key one.

Closes #386.