Fix: `h-geo` and `h-adr` are not child cards
This was an incorrect reading of the spec - the `h-adr` and `h-geo`
classes sit on the same element as `p-location`, otherwise we don't
parse it correctly.

This means we need to add extra validation to ensure that the `location` field is
set, and that the `location` card is of type `h-adr` or `h-geo`
respectively.

As part of #395.