Wrap `p-description` in a <div> not a <p>
`p-description` could have its own markup embedded, so it means that if
we wrap it within a <p>, that <p> will break as soon as Hugo injects
another <p> for the actual content.

Wrapping it in a <div> instead saves us from this issue.

Note that the `p-description` is explicitly a _plain_ data type, not one
with embedded content (that would be an `e-description`), we need to
make sure that it renders correctly when Hugo puts the content in, so
that content is nicely formed for a browser.

As part of #395.