WIP: Allow events to have an embedded h-card
It turns out that this was a bit of an oversight of mine - as called out
by Chris Emerson on Twitter, I hadn't named Ludorati cafe as our
destination.

Then, looking through the Microformats wiki, it appears that to embed
the details of a place (i.e. add a `p-name`) you need an `h-card`.

We therefore need to extend our code to verify that there is an hcard in
the `location` field, and then that it's valid.