Allow `HasField` to not fail if field is not found
For instance, our `EventsPagesHaveValidHevent` is configured to always
check for a `p-latitude`, but we want it to not fail if that field
doesn't exist.

This requires extending it with the ability to turn it on/off, with
default being that it does throw an error if the field is not present.

Note that `fail_if_field_not_found` returns `self` to allow chaining of
methods - mainly so we can use it in the constructor of
`EventsPagesHaveValidHevent` without needing to pull them into separate
variables or to override functionality through our constructor.