Replace `builder` image with multi-stage build
After moving from a complicated Node + Jekyll setup, which required a
heavy `builder` image to track all the dependencies, to Hugo, which is a
single binary, having a separate `builder` image is no longer required.
Now, we just need to pull down the hugo-extended build and we're off.

This means that as part of #210, we can use a multi-stage Docker build,
as we no longer care about the intermediate builder.

This also gives us the opportunity to outsource the existing builder to
the public Docker image created by klakegg, which reduces the
maintenance requirement in this project for upgrades of Hugo.

Unfortunately we can't use the `-onbuild` image, as we need to install
Git onto the image, as well as make the output HTML minified, but maybe
in the future we can!

Closes #210.