Fix: Render RSS/JSON feed content correctly
As noticed in #514, the RSS files were incorrectly formatted. This
was tracked down to when I had massively mis-calculated how `.Render`
works in Hugo.

This is because `.Render` looks for the same output format, which looks
for `content.html` not `content.xml` or `content.jsonfeed.json`.

The solution is to have a Hugo partial (to reduce duplication) in which
to store the HTML render, and then have that called by the custom output
format.

I then needed to make sure that the custom partial is used if one
exists, and fall back to just rendering `.Content` as that's default for
many page types.

Additionally, update `.Render` post with details of what went wrong, so
others are aware in the future.

Closes #514.