Allow broken links to not break the build
As noted in #330, it can be a pain to have a build fail when a link,
completely irrelevant to the changes being added, does not pass
HTMLProofer's `LinkCheck`.

Therefore, it'd be pretty useful to allow the `LinkCheck` to fail,
allowing merging changes that have unrelated fails.

Also, as this is a separate Rake task and CI stage, this means it can be
more easily separated.

Note that we need to remove our custom `Check`s from our `require`s,
otherwise they'll run twice - first on `test:html_proofer` and once on
`test:links`.

Closes #330.