Set publish dates correctly for historical posts
As a follow-up to 3177ae3f, let's actually set these dates correctly.

We can pick these up from the Git commit dates, and makes it more
representative of the actual datetime that these posts were published.

Unfortunately there are a number of posts that were published on the
wrong day, so instead of updating them (and adding an alias to prevent
permalink breakage) we'll just leave them as-is.

Closes #531.