Set publish dates correctly for incorrect historical posts
This is a follow-up to #531, in which I discovered lots of incorrectly
dated posts. This is correcting the dates, using Hugo aliases to ensure
that we do not break any permalinks.

We need to make sure that Kwalify knows about the `aliases` key
otherwise it'll flag our posts as not well-formed.

Because the permalinks don't exist in the sitemap, we need to amend our
`test:permalinks` Rake task to be able to verify that the alias exists
as well as the main link.

As we're now using aliases, HTMLProofer's Checks (at least for posts)
need to support the fact that a file in the posts directory could be an
alias, rather than a post.

Finally, rename the files, as they main permalink has changed. And as
they've changed, `ref`s need to be updated in line with this change,
too.

Closes #533.