Always show last updated on posts
It's almost always going to be different than the author date, and we
want to be showing the metadata for others.

This is important because there are a number of posts that would've been
better as having identical publish and update dates (i.e. back-posted)
but this has blocked it.

In that case, we would not see both the `dt-published` and the
`dt-updated` which then breaks the HTMLProofer `Check`s.

Closes #513.