Add visualisation of posting frequency
As part of #534, I want to make it easier to visualise how often I post
content, of all content types.

For now, look at it in terms of:

- per hour
- per day
- per day of the week
- per month

And will see if there are any other interesting pieces of data I want to
present in the future.

This requires generating the actual counts of data, which I can do using
a custom output format's template, and then render it using Chart.JS.

Plus, add an announcement post.

Closes #534.