Add p-author to h-feed entries
To make sure that h-feed parsers can see the author's information, as
with the other posts on the h-feed, add an inlined h-card with the basic
p-author needed to render in most Indie readers.

Mark this as hidden (by CSS) as it doesn't need to be displayed to the
user.

Closes #497.