Combine Indie post types into Microformats2 content type
As part of jvt.me#572, I wanted to experiment with having all the Indie
content types (i.e. reply, bookmark, like).

This is twofold - firstly it was driven from the work to build my
Micropub endpoint, as it won't need some logic in there to determine the
correct content type to create within my site. But it was also to remove
the weird state where I was partly writing Microformats2-like properties
for my content, but also not very, because it was my own setup around it
as i.e. I didn't want to have to worry about doing that mapping.

This required quite a few changes, to reformat the existing content into
the new Microformats2 format.

We remove schema validation for the new content type as we want to be
doing that validation through the Micropub endpoint rather than in this
site.

Finally, we move the RSVP calendar to a custom output format, as we
can't use it under the RSVPs content type, as that doesn't exist any
more.

Additionally, resolve a bug with how datetimes are rendered.

Closes #572, #500.