Add support for Micropub
As part of creating more types of content, I want to be able to easily
use a Micropub server to push new content to this website.

However, before I can start doing that, I need to publicise my Micropub
endpoint. This will be done, as per the Microsub support, using an HTML
`link` element.

As part of this, also mark my `www-api` as a domain not to check within
HTMLProofer, as this will almost certainly error, as these URLs
shouldn't be hit!