Add a `kind` hint to MF2 content
To make it easier to render the correct content type, as well as have
links such as i.e. /kind/replies/, we want to have the `kind` taxonomy
set up for our content.

This makes it possible to strip some painful duplication that was
introduced as part of #572, while also making it an easier experience
for determining what content type an MF2 entry is.

As part of this, we're also wanting to pluralise the content type, as it
makes the URLs more readable in line with what the content structure has
them named as.

Note that we still need to specify the existing taxonomies we have,
otherwise they won't render!

We also need to ad the `kind` hint to any existing content, so it'll
render properly.

Closes #689.