Use the NodeJS Netlify CLI, not `netlifyctl`
As `netlifyctl` has been deprecated in favour of the Node.JS CLI[0]
(after the Node.JS CLI was deprecated for `netlifyctl`) we should move
to using the new, proffered tool.

This is a shame, as `netlifyctl` was an awesome Go CLI that took zero
time to download.

However, it took a while to deploy - maybe this new version will be
quicker?

[0]: https://www.netlify.com/blog/2018/09/10/netlify-cli-2.0-now-in-beta/#our-cli-journey