Fix: Don't re-wrap /mf2/ entries in an h-entry
This was found when trying to parse my h-feed for post-deploy webmention
support, which made me realise that each `h-entry` had a `h-entry`
child with the actual data in it.

This will also be an issue where the /mf2/ entry may not be an `h-entry`
so would be incorrectly set up.