Render page titles for /mf2/ entries
A long-term issue has been that since adding /mf2/ entries, we've not
been rendering their page titles. This also means that /search/ was
populated with a load of empty unclickable entries which isn't ideal.

Note that we need to add a `trim` in here as otherwise newlines are
being rendered and breaking our JSON file. We also need to make sure
that we use `{{-` and `-}}` to trim any whitespace emitted from the
tags.

Closes #712.