Fix: Allow /mf2/ entries that aren't `h-entry` to not be HTMLProofer'd
When introduced in #595, I had set up the ability to handle the new type
in the theme, but hadn't considered that HTMLProofer looks for an
`h-entry` in /mf2/ which will no longer exist.

This needs to be amended as part of follow-up issue #729, but for now we
can unblock it with the ability to ignore anything that doesn't contain
a `h-entry`.

Plus add a `pending` test which will flag up an implementation gap.