Add default `kind` for /posts/
Now Hugo supports cascading of properties into a content type, it'd be
useful to use this to add the `kind` taxonomy for posts, so we can
easily see the breakdown per type.

Semantically these are called articles in the IndieWeb community, so it
makes sense to call them that here, too, as it fits better than "post"
which is a bit more vague.

We also need to make sure that the new `_index.md` for the cascading
data isn't validated by Kwalify, as it won't be valid!

As part of #733.