Remove time precision for copyright date
It appears that this is the cause of incredibly long (and large) deploys
- the date is incredibly precise, which means it's always different
across (almost) all HTML files in the built site. This

With this now only being set to be the year, we'll have a tonne fewer
differences, and Netlify should only selectively upload the changes that
are actually new, which will massively speed up deployments.