Restructure /mf2/ permalinks for citability
To make the URLs more citable in other forms i.e. Twitter, it'd be more
useful to have them as shorter URLs, as well as including some
information about the date, so they're easier to sort / determine when
they are from.

This requires amending our existing /mf2/ entries, which involves a few
more changes in some files than others due to it being done with the
Ruby pretty-printer opposed to the Java one.

Closes #702.