Silence warnings for `eventscalendar` templates
Because the site is looking up the calendar templates for both the
individual event pages, and their overarching events, we're getting a
tonne of warnings.

This is extra noise and hides other warnings in the logs, so can be
gamed by creating these empty files which will be noticed by Hugo but
not render anything.