Fix: render <blockquote> better on mobile
As mentioned in #707, the rendering of `<blockquote>`s didn't quite work
on smaller devices, as they were either not rendering for the full width
of the page (as each `<p>` was `max-width`'d themselves) or would be
indented too far, reducing the screen width.

This ensures that the `<p>`s within the `<blockquote>` fits nicely and
uses space on mobile more carefully.

Closes #707.