Make URLs case sensitive
It'd be super useful to have case-sensitive URLs for the /mf2/ entries,
so we don't reduce the utility of generating the URLs with capital
letters.

Additionally, we want to add a smoke test for case sensitivity to
protect this functionality in my site in the future. This will validate
that a given /mf2/ entry has both case-sensitive and case-insensitive
URLs available.

Via steps in https://www.jvt.me/posts/2019/11/10/hugo-case-sensitive-urls/

As part of #781.