Replace hardcoded menu with Hugo menu
To make it easier to change the menus on this site, use Hugo's new Menu
functionality[0] to manage this, instead of hardcoding it.

Note that we need the weighting to make them appear in the right order,
and we space out the numbers so they're a bit easier to insert things
between them in the future.

Closes #785.

[0]: https://gohugo.io/content-management/menus/