Validate JSON Feed is well-formed in the pipeline
We've had a number of issues in the past where a change has broken the
JSON feed in production and we haven't known for some time.

Instead of just checking whether it's valid JSON, we can utilise the
JSON Schema (http://json.schemastore.org/feed) to validate that the
whole feed is valid.

Closes #726.