Render inline code snippets for notes
To make it nicer to read notes that contain backticks (that denote
inline code snippets) we should actually wrap them in a `<code>` block.

This is a little nicer if we use an intermediary variable and output it
finally, sanitised by `safeHTML`, rather than doing it at each point.

If a link appears in this `<code>`, it should be ignored by HTML-Proofer
as `<code>`s are not checked.

Plus, make links that are embedded in inline code snippets (i.e. as part
of notes that have URLs autolinked) more readable, we need to make the
contrast better by introducing a lighter colour for the `<a>`s.