Don't check external link resolution in the pipeline
To save a couple of minutes of execution time, we can remove the
optional link checking, because it's not breaking the build anyway, so
is purely for information. If we want to determine if there are any
broken links, we are doing that manually anyway, so we no longer need it
in the pipeline.