Don't re-install Gem dependencies for each build
Instead of re-installing all of the dependencies each time, we can
build a base image with all dependencies pre-installed.

We can do this for both the Hugo image (which we add timezone info on
top of) and the Ruby image (which we install our Gem dependencies).
This then can be re-used from our multi-stage build, but removes a
little of the extra build time, at the cost of that being performed
up-front.