Add custom text for bookmark syndication to Twitter via Brid.gy
As per https://brid.gy/about#silo-content we can use a custom class to
render special output for Brid.dy to syndicate a better message to
Twitter.

We need to make sure to hide this with CSS so it's not displayed, as the
existing output is better for a human person to read.