Only run Rake tests on branches
When pushing straight to master, via Micropub, any validation of the
content should be done before it reaches GitLab, and therefore I should
have confidence it will not break the site's build, or output.

(If it does, that's an issue that needs to be resolved within the
Micropub server. Additionally, if this did break master, it'd be very
difficult for me to resolve it from i.e. mobile.)

Therefore, master builds are slowed down by ~4 minutes for no added
benefit, which means publishing new content via Micropub is slower, and
reduces the real-time feel of my site and Social Web interactions.

Additionally when pushing via the CLI, I should be checking it myself -
and if not, that's on me.

However, on branches I am often adding new articles, or making larger
functional changes to the site which definitely should be validated with
HTMLProofer to ensure that the content is correct, and any changes to
the site do not break my existing functional expectations.

Therefore, we can remove running HTMLProofer, or any other Ruby checks,
as part of the build on master, and instead verify these only on
branches.