Remove GitLab deployments to Netlify
As part of an experiment to use Netlify to build+deploy the site itself,
I need to remove the ability for GitLab CI to deploy over the site,
while Netlify is doing it.

We'll still build the site and create a packaged Docker container for
it, but we won't actually use it.