Use GitLab CI again for Production deploys
Using GitLab CI over Netlify increases the build time by ~1 minute, but
it means that I can avoid the cost implications of Netlify (about ~30
builds have incurred 18 minutes - this will lead to the 300 build minute
limit fairly quickly) as well as not requiring any changes in
/post-deploy to handle Netlify deploy webhooks.

We can also remove the caching for `netlify.toml` as we're going to have
it available for `netlifyctl` as we're checking out the repo for the
build.

We can avoid pushing the Docker image, as we don't actually really need
it in this case, although it's a shame that we'll not have a way to
see what's in Prod.