Add correct markup for Brid.gy quote tweets
As noticed with #864, the markup for a quote tweet with Brid.gy isn't
quite right.

According to gRegor, all we should need to do is add the link to the
post in the `e-content` (as per
https://gregorlove.com/2019/12/i-was-going-into-tosche/). However,
because the `e-content` is separately generated, this isn't working.

We can put that into a separate partial so as not to duplicate code.

Closes #864.