Allow an event's locality to not be present in RSVPs
As found with RSVPing to
https://events.indieweb.org/2020/02/homebrew-website-club-nottingham-8IgcYeAQhIKX,
there is no `locality` exposed in the event metadata. This means that
the site fails to build as it receives a `null` in the field, and when
trying to `index` into it, errors.

That needs to be fixed in www-api separately, as well as handled more
gracefully in the theme.