Add like/reply/repost-context for Twitter interactions
Although I am trying to move to more IndieWeb-only interactions, I still
heavily interact with folks on Twitter.

However, this makes it a bit more difficult for folks reading my social
feed, as they end up with no idea what's going on other than "I like
this person's tweet".

As part of #347, I want to support rendering this
like/reply/repost-context for any Twitter interactions, so folks reading
the feed can see what I'm actually interacting with.

This is added to likes, replies and reposts of Twitter interactions, and
is aware of whether the post I'm interacting with is a quote tweet or a
reply.

The new `context` property is expected to a parsed Microformats2 JSON
response, such as the expected format from Granary.

Unfortunately we don't always get an `alt` tag, so we need to put
something in to tell screen readers that they're unfortunately not
getting anything helpful.

Closes #347.